Add files via upload
This is what I have so far.  There are two timeseries plots of medicare payment and total discharges over time; the third is the most common DRGs over time and the last is the total discharges by state (a bar chart which I don't know if Jeff did but I already had made the code so I thought I would add).  For the first two, I want to modify the plots to include data by region, which would provide a richer plot.